Async issues remove checkpermission fron Global class
